mobile_app/backend/requirements.txt
# arXiv论文助手 - 后端依赖

# FastAPI框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 环境配置
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 网络请求
aiohttp==3.9.1
httpx==0.25.1

# 数据处理
pandas==2.1.4
numpy==1.26.2

# arXiv爬虫
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10

# 推送通知
pyfcm==1.5.4
apns2==0.7.1

# 缓存
redis==5.0.1
aioredis==2.0.1

# 任务队列
celery==5.3.4
flower==2.0.1

# 邮件
python-email-validator==2.1.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# 开发工具
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# 文档
mkdocs==1.5.3
mkdocs-material==9.5.3

# 监控
sentry-sdk==1.38.0
prometheus-client==0.19.0

# 其他
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.3

# 生产环境
gunicorn==21.2.0
uvloop==0.19.0